+++ spec/decoding/decoders/hash_spec.rb
@@ -20,7 +20,13 @@ module Decoding
 
       it "fails when given something other than an hash" do
         hash = Hash.new(Decoders.string, Decoders.integer)
-        expect(hash.call(true)).to eql(Result.err("expected Hash, got TrueClass"))
+        expect(hash.call(true)).to eql(Result.err(Failure.new("expected Hash, got TrueClass")))
+      end
+
+      it "fails when a key cannot be decoded" do
+        hash = Hash.new(Decoders.integer, Decoders.string)
+        expect(hash.call({ "foo" => "bar" }))
+          .to eql(Result.err(Failure.new("error decoding key \"foo\": expected Integer, got String")))
       end
     end
   end