+++ spec/decoding/result_spec.rb
@@ -84,7 +84,7 @@ module Decoding
       expect(Result.err("error").and_then { Result.ok(_1 * 2) }).to eql(Result.err("error"))
     end
 
-    it 'supports pattern matching on results' do
+    it "supports pattern matching on results" do
       case Result.ok(5)
       in Decoding::Ok(n)
         expect(n).to be(5)