+++ spec/spec_helper.rb
@@ -1,5 +1,14 @@
 # frozen_string_literal: true
 
+if ENV["COVERAGE"]
+  require "simplecov"
+  SimpleCov.start do
+    add_filter "/spec/"
+    enable_coverage :branch
+    minimum_coverage line: 100, branch: 100
+  end
+end
+
 require "decoding"
 
 RSpec.configure do |config|