+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 
-## [0.2.3]
+* Ensure different decoders use the same error values
+* Turn exceptions in decoder blocks into errors
+* Use immutable failure values as errors
+
+## [0.2.3] - 2025-10-25
 
 * Implement `Decoding::Result#deconstruct` to support pattern matching on result values.